# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: AIPC Platform API
  description: 'AIPC Platform REST API for edge AI computing platform.


    Provides unified access to:

    - AI Runtime (model management, inference)

    - Event Bus (pub/sub messaging)

    - Device Control (camera, PTZ, GPIO)

    - App Manager (container lifecycle)

    - Stream (HLS video streaming)

    '
  version: 1.0.0
  contact:
    name: AIPC Platform Team
    email: support@camthink.ai
servers:
- url: http://localhost:8080/api/v1
  description: Development server
- url: http://{host}:8080/api/v1
  description: Production server
  variables:
    host:
      default: localhost
security:
- BearerAuth: []
...
paths:
...
  /files:
    get:
      tags:
      - Files
      summary: List directory contents
      operationId: listFiles
      parameters:
      - name: path
        in: query
        schema:
          type: string
          default: /data/aipc
      responses:
        '200':
          description: Directory listing
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/APIResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        path:
                          type: string
                        files:
                          type: array
                          items:
                            type: object
                            properties:
                              name:
                                type: string
                              path:
                                type: string
                              is_dir:
                                type: boolean
                              size:
                                type: integer
                              mod_time:
                                type: string
                              mode:
                                type: string
        '403':
          description: Access denied (path outside allowed roots)
    delete:
      tags:
      - Files
      summary: Delete file or directory
      operationId: deleteFile
      parameters:
      - name: path
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: File deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
        '403':
          description: Access denied
...
  /files/mkdir:
    post:
      tags:
      - Files
      summary: Create directory
      operationId: makeDir
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - path
              properties:
                path:
                  type: string
      responses:
        '200':
          description: Directory created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
...
  /dev/projects/{id}/files:
    get:
      tags: [Dev]
      summary: List project files
      operationId: listProjectFiles
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: path
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of files
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/APIResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        files:
                          type: array
                          items:
                            type: object
                            properties:
                              name:
                                type: string
                              path:
                                type: string
                              is_dir:
                                type: boolean
                              size:
                                type: integer
  /dev/projects/{id}/file:
    get:
      tags: [Dev]
      summary: Get file content
      operationId: getProjectFileContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File content
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/APIResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        path:
                          type: string
                        content:
                          type: string
    post:
      tags: [Dev]
      summary: Save file content
      operationId: saveProjectFileContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path, content]
              properties:
                path:
                  type: string
                content:
                  type: string
      responses:
        '200':
          description: File saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  schemas:
    APIResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: Success
        data:
          type: object
          description: Response payload
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    ErrorDetail:
      type: object
      properties:
        type:
          type: string
        detail:
          type: string

    # ==================== Store Schemas ====================
...